Update initial episode number to test for
Harmontown didn't really start their official video stream until ep. 126. ep. 117 was the first test, and then there's a gap in video. As matching episode numbers, recording dates and air dates is tricky, the easiest solution is to just bump up the initial video episode to ep. 126 in the script and leave the downloading of ep. 117 to the subscriber.